scripts/requirements.txt
psycopg2-binary==2.9.9
watchdog==3.0.0 